Implemented spring-Security contracts and created the classes for registration and authentication

+++ src/main/java/com/ayush/ayush/repository/SellerRepositoryJpa.java
@@ -3,5 +3,5 @@ package com.ayush.ayush.repository;
 import com.ayush.ayush.model.Seller;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SellerRepositoryJpa extends JpaRepository<Seller,Integer> {
+public interface SellerRepositoryJpa extends JpaRepository<Seller,Long> {
 }
